+++ src/formatter/step_definition_snippet_builder/javascript_snippet_syntax_spec.js
@@ -8,19 +8,21 @@ describe('JavascriptSnippetSyntax', function() {
       })
 
       it('returns the proper snippet', function() {
-        const actual = this.syntax.build(
-          'functionName',
-          'pattern',
-          ['arg1', 'arg2', 'callback'],
-          'comment'
-        )
+        const actual = this.syntax.build({
+          comment: 'comment',
+          functionName: 'functionName',
+          generatedExpressions: [
+            {
+              source: 'pattern',
+              parameterNames: ['arg1', 'arg2']
+            }
+          ],
+          stepParameterNames: []
+        })
         const expected =
-          "functionName('pattern', function (arg1, arg2, callback) {" +
-          '\n' +
-          '  // comment' +
-          '\n' +
-          "  callback(null, 'pending');" +
-          '\n' +
+          "functionName('pattern', function (arg1, arg2, callback) {\n" +
+          '  // comment\n' +
+          "  callback(null, 'pending');\n" +
           '});'
         expect(actual).to.eql(expected)
       })
@@ -32,19 +34,21 @@ describe('JavascriptSnippetSyntax', function() {
       })
 
       it('returns the proper snippet', function() {
-        const actual = this.syntax.build(
-          'functionName',
-          'pattern',
-          ['arg1', 'arg2', 'callback'],
-          'comment'
-        )
+        const actual = this.syntax.build({
+          comment: 'comment',
+          functionName: 'functionName',
+          generatedExpressions: [
+            {
+              source: 'pattern',
+              parameterNames: ['arg1', 'arg2']
+            }
+          ],
+          stepParameterNames: []
+        })
         const expected =
-          "functionName('pattern', function *(arg1, arg2) {" +
-          '\n' +
-          '  // comment' +
-          '\n' +
-          "  return 'pending';" +
-          '\n' +
+          "functionName('pattern', function *(arg1, arg2) {\n" +
+          '  // comment\n' +
+          "  return 'pending';\n" +
           '});'
         expect(actual).to.eql(expected)
       })
@@ -56,19 +60,21 @@ describe('JavascriptSnippetSyntax', function() {
       })
 
       it('returns the proper snippet', function() {
-        const actual = this.syntax.build(
-          'functionName',
-          'pattern',
-          ['arg1', 'arg2', 'callback'],
-          'comment'
-        )
+        const actual = this.syntax.build({
+          comment: 'comment',
+          functionName: 'functionName',
+          generatedExpressions: [
+            {
+              source: 'pattern',
+              parameterNames: ['arg1', 'arg2']
+            }
+          ],
+          stepParameterNames: []
+        })
         const expected =
-          "functionName('pattern', function (arg1, arg2) {" +
-          '\n' +
-          '  // comment' +
-          '\n' +
-          "  return 'pending';" +
-          '\n' +
+          "functionName('pattern', function (arg1, arg2) {\n" +
+          '  // comment\n' +
+          "  return 'pending';\n" +
           '});'
         expect(actual).to.eql(expected)
       })
@@ -80,19 +86,21 @@ describe('JavascriptSnippetSyntax', function() {
       })
 
       it('returns the proper snippet', function() {
-        const actual = this.syntax.build(
-          'functionName',
-          'pattern',
-          ['arg1', 'arg2', 'callback'],
-          'comment'
-        )
+        const actual = this.syntax.build({
+          comment: 'comment',
+          functionName: 'functionName',
+          generatedExpressions: [
+            {
+              source: 'pattern',
+              parameterNames: ['arg1', 'arg2']
+            }
+          ],
+          stepParameterNames: []
+        })
         const expected =
-          "functionName('pattern', function (arg1, arg2) {" +
-          '\n' +
-          '  // comment' +
-          '\n' +
-          "  return 'pending';" +
-          '\n' +
+          "functionName('pattern', function (arg1, arg2) {\n" +
+          '  // comment\n' +
+          "  return 'pending';\n" +
           '});'
         expect(actual).to.eql(expected)
       })
@@ -104,19 +112,57 @@ describe('JavascriptSnippetSyntax', function() {
       })
 
       it('returns the proper snippet', function() {
-        const actual = this.syntax.build(
-          'functionName',
-          "pattern'",
-          ['arg1', 'arg2', 'callback'],
-          'comment'
-        )
+        const actual = this.syntax.build({
+          comment: 'comment',
+          functionName: 'functionName',
+          generatedExpressions: [
+            {
+              source: "pattern'",
+              parameterNames: ['arg1', 'arg2']
+            }
+          ],
+          stepParameterNames: []
+        })
         const expected =
-          "functionName('pattern\\'', function (arg1, arg2) {" +
-          '\n' +
-          '  // comment' +
-          '\n' +
-          "  return 'pending';" +
-          '\n' +
+          "functionName('pattern\\'', function (arg1, arg2) {\n" +
+          '  // comment\n' +
+          "  return 'pending';\n" +
+          '});'
+        expect(actual).to.eql(expected)
+      })
+    })
+
+    describe('multiple patterns', function() {
+      beforeEach(function() {
+        this.syntax = new JavascriptSnippetSyntax('synchronous')
+      })
+
+      it('returns the snippet with the other choices commented out', function() {
+        const actual = this.syntax.build({
+          comment: 'comment',
+          functionName: 'functionName',
+          generatedExpressions: [
+            {
+              parameterNames: ['argA', 'argB'],
+              source: 'pattern1'
+            },
+            {
+              parameterNames: ['argC', 'argD'],
+              source: 'pattern2'
+            },
+            {
+              parameterNames: ['argE', 'argF'],
+              source: 'pattern3'
+            }
+          ],
+          stepParameterNames: []
+        })
+        const expected =
+          "functionName('pattern1', function (argA, argB) {\n" +
+          "// functionName('pattern2', function (argC, argD) {\n" +
+          "// functionName('pattern3', function (argE, argF) {\n" +
+          '  // comment\n' +
+          "  return 'pending';\n" +
           '});'
         expect(actual).to.eql(expected)
       })